PN-2712 - finally ended the split on pf-webapp
