+++ packages/pn-personafisica-webapp/src/pages/__test__/Notifiche.page.test.tsx
@@ -1,13 +1,13 @@
+import { 
+  apiOutcomeTestHelper, AppResponseMessage, formatToTimezoneString, getNextDay, 
+  ResponseEventDispatcher, tenYearsAgo, today 
+} from '@pagopa-pn/pn-commons';
 import { act, fireEvent, RenderResult, screen, waitFor, within } from '@testing-library/react';
-import * as redux from 'react-redux';
-import { formatToTimezoneString, getNextDay, apiOutcomeTestHelper, tenYearsAgo, today, ResponseEventDispatcher, AppResponseMessage } from '@pagopa-pn/pn-commons';
 
-import { axe, render } from '../../__test__/test-utils';
-import * as actions from '../../redux/dashboard/actions';
-import * as hooks from '../../redux/hooks';
-import { notificationsToFe } from '../../redux/dashboard/__test__/test-utils';
-import Notifiche from '../Notifiche.page';
 import { NotificationsApi } from '../../api/notifications/Notifications.api';
+import { render } from '../../__test__/test-utils';
+import Notifiche from '../Notifiche.page';
+import { doPrepareTestScenario } from './Notifiche.page.test-utils';
 
 jest.mock('react-i18next', () => ({
   // this mock makes sure any components using the translate hook can use it without a warning being shown
@@ -33,48 +33,13 @@ jest.mock('react-i18next', () => ({
 describe('Notifiche Page - with notifications', () => {
   let result: RenderResult | undefined;
   let mockDispatchFn: jest.Mock;
-
-  const mockActionFn = jest.fn();
+  let mockActionFn: jest.Mock;
 
   beforeEach(async () => {
-    mockDispatchFn = jest.fn(() => ({
-      then: () => Promise.resolve(),
-    }));
-    // mock app selector
-    const spy = jest.spyOn(hooks, 'useAppSelector');
-    spy
-      .mockReturnValueOnce({
-        notifications: notificationsToFe.resultsPage,
-        filters: {
-          startDate: formatToTimezoneString(tenYearsAgo),
-          endDate: formatToTimezoneString(today),
-          recipientId: '',
-          status: '',
-          subjectRegExp: '',
-        },
-        sort: {
-          orderBy: '',
-          order: 'asc',
-        },
-        pagination: {
-          nextPagesKey: ['mocked-page-key-1', 'mocked-page-key-2', 'mocked-page-key-3'],
-          size: 10,
-          page: 0,
-          moreResult: true,
-        },
-      })
-      .mockReturnValueOnce({ delegators: [] })
-      .mockReturnValueOnce({ legalDomicile: [] });
-    // mock action
-    const actionSpy = jest.spyOn(actions, 'getReceivedNotifications');
-    actionSpy.mockImplementation(mockActionFn);
-    // mock dispatch
-    const useDispatchSpy = jest.spyOn(redux, 'useDispatch');
-    useDispatchSpy.mockReturnValue(mockDispatchFn as any);
-    // render component
-    await act(async () => {
-      result = render(<Notifiche />);
-    });
+    const scenario = await doPrepareTestScenario();
+    result = scenario.result;
+    mockDispatchFn = scenario.mockDispatchFn;
+    mockActionFn = scenario.mockActionFn;
   });
 
   afterEach(() => {
@@ -143,15 +108,6 @@ describe('Notifiche Page - with notifications', () => {
       });
     });
   });
-
-  it.skip('does not have basic accessibility issues', async () => {
-    if (result) {
-      const res = await axe(result.container);
-      expect(res).toHaveNoViolations();
-    } else {
-      fail("render() returned undefined!");
-    }
-  }, 15000);
 });
